Create robotics_array.rs
In this revised script, we added a machine learning system and a natural language processor. We also added code to train the machine learning system using the integrated matrix, use it to make a prediction, parse a command using the natural language processor, and execute the command using the robotics library. Note that this code is just an example and may not run as is, as it depends on the definitions